主数据管理模块完成，购物车模块完成

+++ G2Catering/G2Catering/Main(主界面)/Ordering(点单)/Controller/ShowGoodsView.h
@@ -11,7 +11,7 @@
 
 @protocol ShowGoodsViewViewDelegate <NSObject>
 
--(void)btnDelegateName:(ShopGoodsDetailModel *)goodCheck number:(int)number type:(int)type;
+-(void)btnDelegateName:(ShopGoodsDetailModel *)goodCheck number:(int)number type:(NSInteger)type;
 
 @end
 @interface ShowGoodsView : UIView
